requirements.txt
flask
PyPDF2
pinecone
python-dotenv